Create HasHealthAnswers concern
This adds a concern which handles common functionality related to having
health answers, which applies to the `Consent`, `ConsentForm` and
`DraftConsent` models.
